+++ CampaignEditor/UserControls/ForecastGrids/SpotWeekGoalsGrid.xaml.cs
@@ -220,8 +220,8 @@ namespace CampaignEditor.UserControls
             var channelMpTuples = _visibleTuples.Where(mpt => mpt.MediaPlan.chid == channel.chid);
 
             int ins = 0;
-            double grp = 0;
-            double budget = 0;
+            decimal grp = 0;
+            decimal budget = 0;
             foreach (var mpTuple in channelMpTuples)
             {
                 foreach (int index in weekIndexes)
@@ -298,8 +298,8 @@ namespace CampaignEditor.UserControls
                                                 .Select(spotSpotGoalsDict => spotSpotGoalsDict.Value));
 
             int ins = 0;
-            double grp = 0;
-            double budget = 0;
+            decimal grp = 0;
+            decimal budget = 0;
             foreach (var spotGoal in channelSpotGoals)
             {
                 ins += spotGoal.Insertations;
@@ -330,8 +330,8 @@ namespace CampaignEditor.UserControls
             var totalSpotGoals = _data[dummyChannel][weekNum][spot];
 
             int ins = 0;
-            double grp = 0;
-            double budget = 0;
+            decimal grp = 0;
+            decimal budget = 0;
 
             var channelSpotGoals = _data.Where(dict => _selectedChannels.Contains(dict.Key)) //  all selected channels
                                         .SelectMany(dict => dict.Value) // use only values
