+++ Database/DTOs/PricelistDTO/PricelistDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.PricelistDTO
 {
     public class PricelistDTO : BaseIdentityPricelistDTO
     {
-        public PricelistDTO(int plid, int clid, string plname, int pltype, int sectbid, int seastbid, bool plactive, float price, float minprice, bool prgcoef, int pltarg, bool use2, int sectbid2, int sectb2st, int sectb2en, int valfrom, int valto, bool mgtype) 
+        public PricelistDTO(int plid, int clid, string plname, int pltype, int sectbid, int seastbid, bool plactive, decimal price, decimal minprice, bool prgcoef, int pltarg, bool use2, int sectbid2, int sectb2st, int sectb2en, int valfrom, int valto, bool mgtype) 
             : base(plid, clid, plname, pltype, sectbid, seastbid, plactive, price, minprice, prgcoef, pltarg, use2, sectbid2, sectb2st, sectb2en, valfrom, valto, mgtype)
         {
         }
