+++ Database/DTOs/PricesDTO/BasePricesDTO.cs
@@ -4,7 +4,7 @@ namespace Database.DTOs.PricesDTO
 {
     public class BasePricesDTO
     {
-        public BasePricesDTO(int plid, string dps, string dpe, float price, bool ispt, string days)
+        public BasePricesDTO(int plid, string dps, string dpe, decimal price, bool ispt, string days)
         {
             this.plid = plid;
             this.dps = dps ?? throw new ArgumentNullException(nameof(dps));
@@ -17,7 +17,7 @@ namespace Database.DTOs.PricesDTO
         public int plid { get; set; }
         public string dps { get; set; }
         public string dpe { get; set; }
-        public float price { get; set; }
+        public decimal price { get; set; }
         public bool ispt { get; set; }
         public string days { get; set; }
     }
