+++ Database/DTOs/SectablesDTO/BaseIdentitySectablesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SectablesDTO
 {
     public class BaseIdentitySectablesDTO
     {
-        public BaseIdentitySectablesDTO(int sctid, int sec, double coef)
+        public BaseIdentitySectablesDTO(int sctid, int sec, decimal coef)
         {
             this.sctid = sctid;
             this.sec = sec;
@@ -12,6 +12,6 @@ namespace Database.DTOs.SectablesDTO
 
         public int sctid { get; set; }
         public int sec { get; set; }
-        public double coef { get; set; }
+        public decimal coef { get; set; }
     }
 }
