+++ Database/DTOs/SectablesDTO/CreateSectablesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SectablesDTO
 {
     public class CreateSectablesDTO : BaseIdentitySectablesDTO
     {
-        public CreateSectablesDTO(int sctid, int sec, double coef) 
+        public CreateSectablesDTO(int sctid, int sec, decimal coef) 
             : base(sctid, sec, coef)
         {
         }
