+++ Database/DTOs/SectablesDTO/SectablesDTO.cs
@@ -3,7 +3,7 @@ namespace Database.DTOs.SectablesDTO
 {
     public class SectablesDTO : BaseIdentitySectablesDTO
     {
-        public SectablesDTO(int sctid, int sec, double coef) 
+        public SectablesDTO(int sctid, int sec, decimal coef) 
             : base(sctid, sec, coef)
         {
         }
