+++ Database/Entities/Sectables.cs
@@ -5,6 +5,6 @@ namespace Database.Entities
     {
         public int sctid { get; set; }
         public int sec { get; set; }
-        public double coef { get; set; }
+        public decimal coef { get; set; }
     }
 }
